Extract tree address prefix decoding
